refactor: migrate.goをdriverパッケージに移動
