+++ backend/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"fmt"
 	"myapp/internal/config"
-	"myapp/internal/infrastructure/persistence/datastore"
+	"myapp/internal/infrastructure/persistence/datastore/driver"
 	"myapp/internal/interface/api/middleware"
 	"myapp/internal/interface/api/router"
 )
 
 func main() {
-	var migrator middleware.DBMigrator = datastore.MustNewMySQLMigrator("file://migrate")
+	var migrator middleware.DBMigrator = driver.MustNewMySQLMigrator("file://migrate")
 	if err := migrator.Migrate(); err != nil {
 		panic(err)
 	}
